atlas-texture-creator/pyproject.toml
[tool.poetry]
name = "atlas-texture-creator"
version = "0.2.0"
description = ""
authors = ["pglaw <philipp.glaw@gmx.de>"]
readme = "README.md"
packages = [{include = "atlas_texture_creator"}]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.1.0"
sqlmodel = "^0.0.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
decorator = "^5.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
